updated README with more info + some renames of files
